+++ src/main/java/dev/minguinho/zeze/domain/common/model/BaseEntity.java
@@ -1,4 +1,4 @@
-package dev.minguinho.zeze.domain.common;
+package dev.minguinho.zeze.domain.common.model;
 
 import java.time.ZonedDateTime;
 import javax.persistence.Column;
