+++ src/main/java/dev/minguinho/zeze/pagecontroller/PageController.java
@@ -1,4 +1,4 @@
-package dev.minguinho.zeze.domain.common;
+package dev.minguinho.zeze.pagecontroller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
